move contract-address test into syscalls_test and remove the module
The module was empty, and mapped to an external type `ContractAddress`.
Better to just move the test unless we have multiple such tests.
