+++ crates/blockifier/src/execution.rs
@@ -1,5 +1,4 @@
 pub mod common_hints;
-pub mod contract_address;
 pub mod contract_class;
 pub mod entry_point;
 pub mod errors;
