--- crates/blockifier/src/execution/contract_address.rs
@@ -1,3 +1,0 @@
-#[cfg(test)]
-#[path = "contract_address_test.rs"]
-pub mod test;
